change panel disconnect
