+++ Result_yolo3_panel/.gitignore
@@ -4,7 +4,8 @@ Result_Prueba/
 Prueba/
 Otros_2/
 Otros/
+Result_Prueba_disconnect
 
-*.jpg
+
 *.h5
 *.pkl
